Refactor of delete() -> added excepetion when ID does not exist

+++ src/main/java/cz/godless/task_management_system/implementation/jdbc/service/UserServiceJdbcImpl.java
@@ -23,7 +23,9 @@ public class UserServiceJdbcImpl implements UserService {
 
     @Override
     public void delete(long id) {
-        repository.delete(id);
+        if (this.get(id) != null) {
+            repository.delete(id);
+        }
     }
 
     @Override
